Moving to conf system - CoordinateGridMixin 01
