+++ src/base/coordinate-grid-mixin.ts
@@ -92,7 +92,7 @@ export class CoordinateGridMixin extends ColorMixin(MarginMixin) {
         this._origX = undefined; // Will hold original scale in case of zoom
         this._xOriginalDomain = undefined;
         this._xAxis = axisBottom(undefined);
-        this._xUnits = units.integers;
+        this._conf.xUnits = units.integers;
         this._xAxisPadding = 0;
         this._xAxisPaddingUnit = timeDay;
         this._xElasticity = false;
@@ -320,53 +320,6 @@ export class CoordinateGridMixin extends ColorMixin(MarginMixin) {
     }
 
     /**
-     * Set or get the xUnits function. The coordinate grid chart uses the xUnits function to calculate
-     * the number of data projections on the x axis such as the number of bars for a bar chart or the
-     * number of dots for a line chart.
-     *
-     * This function is expected to return a Javascript array of all data points on the x axis, or
-     * the number of points on the axis. d3 time range functions [d3.timeDays, d3.timeMonths, and
-     * d3.timeYears](https://github.com/d3/d3-time/blob/master/README.md#intervals) are all valid
-     * xUnits functions.
-     *
-     * dc.js also provides a few units function, see the {@link units Units Namespace} for
-     * a list of built-in units functions.
-     *
-     * Note that as of dc.js 3.0, `units.ordinal` is not a real function, because it is not
-     * possible to define this function compliant with the d3 range functions. It was already a
-     * magic value which caused charts to behave differently, and now it is completely so.
-     * @example
-     * // set x units to count days
-     * chart.xUnits(d3.timeDays);
-     * // set x units to count months
-     * chart.xUnits(d3.timeMonths);
-     *
-     * // A custom xUnits function can be used as long as it follows the following interface:
-     * // units in integer
-     * function(start, end) {
-     *      // simply calculates how many integers in the domain
-     *      return Math.abs(end - start);
-     * }
-     *
-     * // fixed units
-     * function(start, end) {
-     *      // be aware using fixed units will disable the focus/zoom ability on the chart
-     *      return 1000;
-     * }
-     * @param {Function} [xUnits=units.integers]
-     * @returns {Function|CoordinateGridMixin}
-     */
-    public xUnits (): Units;
-    public xUnits (xUnits: Units): this;
-    public xUnits (xUnits?) {
-        if (!arguments.length) {
-            return this._xUnits;
-        }
-        this._xUnits = xUnits;
-        return this;
-    }
-
-    /**
      * Set or get the x axis used by a particular coordinate grid chart instance. This function is most
      * useful when x axis customization is required. The x axis in dc.js is an instance of a
      * {@link https://github.com/d3/d3-axis/blob/master/README.md#axisBottom d3 bottom axis object};
@@ -469,7 +422,7 @@ export class CoordinateGridMixin extends ColorMixin(MarginMixin) {
                 this._unitCount = this.x().domain().length;
             } else {
                 const [first, second] = this.x().domain();
-                const unitCount = this.xUnits()(first, second);
+                const unitCount = this._conf.xUnits(first, second);
 
                 // Sometimes xUnits() may return an array while sometimes directly the count
                 this._unitCount = unitCount instanceof Array ? unitCount.length : unitCount;
@@ -511,7 +464,7 @@ export class CoordinateGridMixin extends ColorMixin(MarginMixin) {
      * @returns {Boolean}
      */
     public isOrdinal (): boolean {
-        return this.xUnits() === units.ordinal;
+        return this._conf.xUnits === units.ordinal;
     }
 
     public _useOuterPadding (): boolean {
