+++ src/base/i-coordinate-grid-mixin-conf.ts
@@ -1,6 +1,7 @@
 import {IMarginMixinConf} from './i-margin-mixin-conf';
 import {IColorMixinConf} from './i-color-mixin-conf';
+import {Units} from '../core/types';
 
 export interface ICoordinateGridMixinConf extends IMarginMixinConf, IColorMixinConf {
-
+    xUnits?: Units;
 }
